airbrush: tmu: isr: fix -Wpointer-to-int-cast
drivers/misc/airbrush/tmu/isr.c:81:18: error: cast to smaller integer type 'u32' (aka 'unsigned int') from 'void *' [-Werror,-Wvoid-pointer-to-int-cast]
        u32 intnc_val = (u32)data;
                        ^~~~~~~~~

According to abc_pcie_irq_handler() @ drivers/mfd/abc-pcie.c,
the value passed into data is a `u64` instead of an
`enum abc_msi_mst_t`. The `intnc_val` is a value read from a 4-byte
hardware register, so that it is casted to a `u32` finally
in ab_tmu_isr_notify.

Bug: 171827315

+++ drivers/misc/airbrush/tmu/isr.c
@@ -78,7 +78,7 @@ static int ab_tmu_isr_notify(struct notifier_block *nb,
 		unsigned long irq, void *data)
 {
 	struct ab_tmu_isr *isr = container_of(nb, struct ab_tmu_isr, nb);
-	u32 intnc_val = (u32)data;
+	u32 intnc_val = (u64)data;
 
 	if (irq == ABC_MSI_AON_INTNC &&
 			(intnc_val & (1 << (isr->irq - ABC_MSI_COUNT))))
